Split AsyncResolver out of Resolver
