+++ src/Grpc.Net.Client/Balancer/Resolver.cs
@@ -48,34 +48,6 @@ namespace Grpc.Net.Client.Balancer
     /// </summary>
     public abstract class Resolver : IDisposable
     {
-        // Internal for testing
-        internal Task _resolveTask = Task.CompletedTask;
-        private Action<ResolverResult>? _listener;
-        private bool _disposed;
-
-        private readonly object _lock = new object();
-        private readonly CancellationTokenSource _cts = new CancellationTokenSource();
-        private readonly ILogger _logger;
-
-        /// <summary>
-        /// Gets the listener.
-        /// </summary>
-        protected Action<ResolverResult> Listener => _listener!;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="Resolver"/>.
-        /// </summary>
-        /// <param name="loggerFactory">The logger factory.</param>
-        protected Resolver(ILoggerFactory loggerFactory)
-        {
-            if (loggerFactory == null)
-            {
-                throw new ArgumentNullException(nameof(loggerFactory));
-            }
-
-            _logger = loggerFactory.CreateLogger<Resolver>();
-        }
-
         /// <summary>
         /// Starts listening to resolver for results with the specified callback. Can only be called once.
         /// <para>
@@ -84,33 +56,7 @@ namespace Grpc.Net.Client.Balancer
         /// </para>
         /// </summary>
         /// <param name="listener">The callback used to receive updates on the target.</param>
-        public void Start(Action<ResolverResult> listener)
-        {
-            if (listener == null)
-            {
-                throw new ArgumentNullException(nameof(listener));
-            }
-
-            if (_listener != null)
-            {
-                throw new InvalidOperationException("Resolver has already been started.");
-            }
-
-            _listener = (result) =>
-            {
-                Log.ResolverResult(_logger, GetType(), result.Status.StatusCode, result.Addresses?.Count ?? 0);
-                listener(result);
-            };
-
-            OnStarted();
-        }
-
-        /// <summary>
-        /// Executes after the resolver starts.
-        /// </summary>
-        protected virtual void OnStarted()
-        {
-        }
+        public abstract void Start(Action<ResolverResult> listener);
 
         /// <summary>
         /// Refresh resolution. Can only be called after <see cref="Start(Action{ResolverResult})"/>.
@@ -118,65 +64,7 @@ namespace Grpc.Net.Client.Balancer
         /// This is only a hint. Implementation takes it as a signal but may not start resolution.
         /// </para>
         /// </summary>
-        public void Refresh()
-        {
-            if (_disposed)
-            {
-                throw new ObjectDisposedException(nameof(DnsResolver));
-            }
-            if (_listener == null)
-            {
-                throw new InvalidOperationException("Resolver hasn't been started.");
-            }
-
-            lock (_lock)
-            {
-                Log.ResolverRefreshRequested(_logger, GetType());
-
-                if (_resolveTask.IsCompleted)
-                {
-                    _resolveTask = ResolveNowAsync(_cts.Token);
-                }
-                else
-                {
-                    Log.ResolverRefreshIgnored(_logger, GetType());
-                }
-            }
-        }
-
-        private async Task ResolveNowAsync(CancellationToken cancellationToken)
-        {
-            try
-            {
-                await ResolveAsync(cancellationToken).ConfigureAwait(false);
-            }
-            catch (OperationCanceledException) when (cancellationToken.IsCancellationRequested)
-            {
-                // Ignore cancellation.
-            }
-            catch (Exception ex)
-            {
-                Log.ResolverRefreshError(_logger, GetType(), ex);
-
-                var status = GrpcProtocolHelpers.CreateStatusFromException("Error refreshing resolver.", ex);
-                Listener(ResolverResult.ForFailure(status));
-            }
-        }
-
-        /// <summary>
-        /// Resolve the target <see cref="Uri"/>. Updated results are passed to the callback
-        /// registered by <see cref="Start(Action{ResolverResult})"/>. Can only be called
-        /// after the resolver has started.
-        /// <para>
-        /// This is only a hint. Implementation takes it as a signal but may not start resolution.
-        /// </para>
-        /// </summary>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>A task.</returns>
-        protected virtual Task ResolveAsync(CancellationToken cancellationToken)
-        {
-            return Task.CompletedTask;
-        }
+        public abstract void Refresh();
 
         /// <summary>
         /// Releases the unmanaged resources used by the <see cref="LoadBalancer"/> and optionally releases
@@ -187,7 +75,6 @@ namespace Grpc.Net.Client.Balancer
         /// </param>
         protected virtual void Dispose(bool disposing)
         {
-            _cts.Cancel();
         }
 
         /// <summary>
@@ -197,43 +84,6 @@ namespace Grpc.Net.Client.Balancer
         {
             Dispose(disposing: true);
             GC.SuppressFinalize(this);
-
-            _disposed = true;
-        }
-
-        internal static class Log
-        {
-            private static readonly Action<ILogger, string, Exception?> _resolverRefreshRequested =
-                LoggerMessage.Define<string>(LogLevel.Trace, new EventId(1, "ResolverRefreshRequested"), "{ResolveType} refresh requested.");
-
-            private static readonly Action<ILogger, string, Exception?> _resolverRefreshIgnored =
-                LoggerMessage.Define<string>(LogLevel.Trace, new EventId(2, "ResolverRefreshIgnored"), "{ResolveType} refresh ignored because resolve is already in progress.");
-
-            private static readonly Action<ILogger, string, Exception?> _resolverRefreshError =
-                LoggerMessage.Define<string>(LogLevel.Error, new EventId(3, "ResolverRefreshError"), "Error refreshing {ResolveType}.");
-
-            private static readonly Action<ILogger, string, StatusCode, int, Exception?> _resolverResult =
-                LoggerMessage.Define<string, StatusCode, int>(LogLevel.Trace, new EventId(4, "ResolverResult"), "{ResolveType} result with status code '{StatusCode}' and {AddressCount} addresses.");
-
-            public static void ResolverRefreshRequested(ILogger logger, Type resolverType)
-            {
-                _resolverRefreshRequested(logger, resolverType.Name, null);
-            }
-
-            public static void ResolverRefreshIgnored(ILogger logger, Type resolverType)
-            {
-                _resolverRefreshIgnored(logger, resolverType.Name, null);
-            }
-
-            public static void ResolverRefreshError(ILogger logger, Type resolverType, Exception ex)
-            {
-                _resolverRefreshError(logger, resolverType.Name, ex);
-            }
-
-            public static void ResolverResult(ILogger logger, Type resolverType, StatusCode statusCode, int addressCount)
-            {
-                _resolverResult(logger, resolverType.Name, statusCode, addressCount, null);
-            }
         }
     }
